feat: add Random Game button on game lists
