+++ app/Community/Controllers/Api/UserGameListApiController.php
@@ -7,6 +7,7 @@ use App\Http\Controller;
 use App\Models\User;
 use App\Models\UserGameListEntry;
 use App\Platform\Actions\BuildGameListAction;
+use App\Platform\Actions\GetRandomGameAction;
 use App\Platform\Enums\GameListType;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -66,5 +67,16 @@ class UserGameListApiController extends Controller
         }
 
         return response()->json(['success' => true]);
+    }
+
+    public function random(UserGameListRequest $request): JsonResponse
+    {
+        $randomGame = (new GetRandomGameAction())->execute(
+            GameListType::UserPlay,
+            user: $request->user(),
+            filters: $request->getFilters(),
+        );
+
+        return response()->json(['gameId' => $randomGame->id]);
     }
 }
