+++ app/Platform/Actions/BuildGameListAction.php
@@ -5,32 +5,24 @@ declare(strict_types=1);
 namespace App\Platform\Actions;
 
 use App\Community\Enums\AwardType;
-use App\Community\Enums\ClaimStatus;
-use App\Community\Enums\TicketState;
 use App\Community\Enums\UserGameListType;
 use App\Data\PaginatedData;
-use App\Models\AchievementSetClaim;
 use App\Models\Game;
-use App\Models\Leaderboard;
 use App\Models\PlayerGame;
-use App\Models\System;
-use App\Models\Ticket;
 use App\Models\User;
+use App\Platform\Concerns\BuildsGameListQueries;
 use App\Platform\Data\GameData;
 use App\Platform\Data\GameListEntryData;
 use App\Platform\Data\PlayerGameData;
-use App\Platform\Enums\AchievementFlag;
-use App\Platform\Enums\GameListSortField;
 use App\Platform\Enums\GameListType;
-use App\Platform\Enums\UnlockMode;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\DB;
-use InvalidArgumentException;
 
 class BuildGameListAction
 {
+    use BuildsGameListQueries;
+
     public function execute(
         GameListType $listType,
         ?User $user,
@@ -127,503 +119,6 @@ class BuildGameListAction
             unfilteredTotal: $unfilteredTotal,
             items: $transformedEntries,
         );
-    }
-
-    /**
-     * @return Builder<Game>
-     */
-    private function buildBaseQuery(GameListType $listType, ?User $user = null): Builder
-    {
-        $query = Game::with(['system'])
-            ->withLastAchievementUpdate()
-            ->addSelect(['GameData.*'])
-            ->addSelect([
-                // Fetch counts here to avoid N+1 query problems.
-
-                'has_active_or_in_review_claims' => AchievementSetClaim::selectRaw('CASE WHEN COUNT(*) > 0 THEN 1 ELSE 0 END')
-                    ->whereColumn('SetClaim.game_id', 'GameData.ID')
-                    ->whereIn('Status', [ClaimStatus::Active, ClaimStatus::InReview])
-                    ->limit(1),
-
-                'num_visible_leaderboards' => Leaderboard::selectRaw('COUNT(*)')
-                    ->whereColumn('LeaderboardDef.GameID', 'GameData.ID')
-                    ->where('LeaderboardDef.DisplayOrder', '>=', 0),
-            ]);
-
-        // Only attempt to fetch the "Open Tickets" column counts if the user
-        // is a dev. Otherwise, skip it.
-        if ($user?->can('develop')) {
-            $query->addSelect([
-                'num_unresolved_tickets' => Ticket::selectRaw('COUNT(*)')
-                    ->join('Achievements', 'Ticket.AchievementID', '=', 'Achievements.ID')
-                    ->whereColumn('Achievements.GameID', 'GameData.ID')
-                    ->where('Achievements.Flags', AchievementFlag::OfficialCore->value)
-                    ->whereIn('Ticket.ReportState', [TicketState::Open, TicketState::Request]),
-            ]);
-        }
-
-        switch ($listType) {
-            case GameListType::AllGames:
-                // Exclude non game systems, inactive systems, and subsets.
-                $validSystemIds = System::active()
-                    ->gameSystems()
-                    ->pluck('ID')
-                    ->all();
-
-                $query
-                    ->whereIn('GameData.ConsoleID', $validSystemIds)
-                    ->where('GameData.Title', 'not like', "%[Subset -%");
-                break;
-
-            case GameListType::UserPlay:
-                $query->whereHas('gameListEntries', function ($query) use ($user) {
-                    $query->where('user_id', $user->id)
-                          ->where('type', UserGameListType::Play);
-                });
-                break;
-
-            // TODO implement these other use cases
-            case GameListType::UserDevelop:
-            case GameListType::System:
-            case GameListType::Hub:
-            case GameListType::DeveloperSets:
-                throw new InvalidArgumentException("List type not implemented");
-            default:
-                throw new InvalidArgumentException("Invalid list type");
-        }
-
-        return $query;
-    }
-
-    /**
-     * @param Builder<Game> $query
-     */
-    private function applyFilters(Builder $query, array $filters, ?User $user = null): void
-    {
-        foreach ($filters as $filterKey => $filterValues) {
-            switch ($filterKey) {
-                /*
-                 * only show games matching a specific game title pattern
-                 */
-                case 'title':
-                    if (!empty($filterValues[0])) {
-                        $query->where('GameData.Title', 'LIKE', '%' . $filterValues[0] . '%');
-                    }
-                    break;
-
-                /*
-                 * only show games matching a specific list of system IDs
-                 */
-                case 'system':
-                    $query->whereHas('system', function (Builder $query) use ($filterValues) {
-                        $query->whereIn('ID', $filterValues);
-                    });
-                    break;
-
-                /*
-                 * only show games matching a specific list of user earned awards
-                 */
-                case 'award':
-                    $this->applyAwardFilter($query, $filterValues, $user);
-                    break;
-
-                /*
-                 * only show games based on whether they have achievements published
-                 */
-                case 'achievementsPublished':
-                    $this->applyAchievementsPublishedFilter($query, $filterValues);
-                    break;
-
-                default:
-                    break;
-            }
-        }
-    }
-
-    /**
-     * @param Builder<Game> $query
-     * @param array{field: string, direction: 'asc'|'desc'} $sort
-     */
-    private function applySorting(Builder $query, array $sort, ?User $user = null): void
-    {
-        if (isset($sort['field']) && GameListSortField::tryFrom($sort['field'])) {
-            $sortDirection = $sort['direction'] ?? 'asc';
-
-            switch ($sort['field']) {
-                /*
-                 * game title, with tagged games placed at the bottom of the list
-                 */
-                case GameListSortField::Title->value:
-                    $query->orderBy('GameData.sort_title', $sortDirection);
-                    break;
-
-                /*
-                 * game system name, by name_short (eg: "A2600", not "Atari 2600")
-                 */
-                case GameListSortField::System->value:
-                    $query
-                        ->join('Console', 'GameData.ConsoleID', '=', 'Console.ID')
-                        ->orderBy('Console.name_short', $sortDirection);
-                    break;
-
-                /*
-                 * count of official achievements associated with the game's core set
-                 */
-                case GameListSortField::AchievementsPublished->value:
-                    $query->orderBy('GameData.achievements_published', $sortDirection);
-                    break;
-
-                /*
-                 * whether or not there are any active or in review claims associated with the game
-                 */
-                case GameListSortField::HasActiveOrInReviewClaims->value:
-                    $query->orderBy('has_active_or_in_review_claims', $sortDirection);
-                    break;
-
-                /*
-                 * count of points from core/official achievements associated with the game's core set
-                 */
-                case GameListSortField::PointsTotal->value:
-                    $query->orderBy('GameData.points_total', $sortDirection);
-                    break;
-
-                /*
-                 * points_weighted / points_total from core/official achievements
-                 * associated with the game's core set
-                 */
-                case GameListSortField::RetroRatio->value:
-                    $query
-                        ->selectRaw(
-                            "CASE
-                                WHEN GameData.points_total = 0 THEN 0
-                                ELSE GameData.TotalTruePoints / GameData.points_total
-                            END AS retro_ratio"
-                        )
-                        ->orderBy('retro_ratio', $sortDirection);
-                    break;
-
-                /*
-                 * when an update was last made to the game's achievement logic
-                 * TODO use updates from the triggers table, achievement logic changes is what players care about
-                 *      and DateModified includes when other stuff changed like titles, descriptions, etc
-                 */
-                case GameListSortField::LastUpdated->value:
-                    $query
-                        ->selectRaw(
-                            "COALESCE(
-                                (SELECT MAX(DateModified) FROM Achievements WHERE Achievements.GameID = GameData.ID),
-                                GameData.Updated
-                            ) AS last_updated"
-                        )
-                        ->orderBy('last_updated', $sortDirection);
-                    break;
-
-                /*
-                 * the game's earliest release date
-                 */
-                case GameListSortField::ReleasedAt->value:
-                    $this->applyReleasedAtSorting($query, $sortDirection);
-                    break;
-
-                /*
-                 * count of all players (softcore and hardcore) for the game
-                 */
-                case GameListSortField::PlayersTotal->value:
-                    $query->orderBy('GameData.players_total', $sortDirection);
-                    break;
-
-                /*
-                 * the game's count of non-hidden leaderboards (order_column >= 0)
-                 */
-                case GameListSortField::NumVisibleLeaderboards->value:
-                    $query->orderBy('num_visible_leaderboards', $sortDirection);
-                    break;
-
-                /*
-                 * the game's count of tickets awaiting resolution
-                 */
-                case GameListSortField::NumUnresolvedTickets->value:
-                    $query->orderBy('num_unresolved_tickets', $sortDirection);
-                    break;
-
-                /*
-                 * the user's progress, ordered by # of achievements earned, on the game
-                 */
-                case GameListSortField::Progress->value:
-                    $this->applyProgressSorting($query, $sortDirection, $user);
-                    break;
-
-                /*
-                 * if we have no idea what the user is trying to sort by, fall back to sorting by title
-                 */
-                default:
-                    $query->orderBy('GameData.sort_title', $sortDirection);
-                    break;
-            }
-        }
-
-        // Default to sorting by title if no valid sort field is provided.
-        // Otherwise, always secondary sort by title.
-        $query->orderBy('GameData.sort_title', 'asc');
-    }
-
-    /**
-     * Sort games on the list by their release dates.
-     * The released_at_granularity value must be taken into account when sorting.
-     * This column indicates the level of precision available for the release date
-     * (eg: year, month, or exact day).
-     *
-     * The sorting logic works as follows:
-     *
-     * 1. If released_at_granularity is set to "year", the release date is normalized
-     *    to the first day of that year (eg: "1985-01-01").
-     * 2. If released_at_granularity is set to "month", the release date is normalized
-     *    to the first day of that month (eg: "1985-05-01").
-     * 3. If no granularity is set, or the granularity is "day", the release date is used as-is.
-     *
-     * This ensures that games with less precise release dates are sorted logically while
-     * maintaining the correct order relative to their peers. Games are then ordered by their
-     * normalized release date, either ascending or descending.
-     *
-     * @param Builder<Game> $query
-     */
-    private function applyReleasedAtSorting(Builder $query, string $sortDirection = 'asc'): void
-    {
-        // We're extra careful here to use functions supported by both MariaDB
-        // and SQLite. This is preferable to altering the query specifically for
-        // SQLite, because if we do so then we can't actually trust any test results.
-        $query
-            ->selectRaw(
-                "GameData.*,
-                CASE
-                    WHEN GameData.released_at_granularity = 'year' THEN
-                        DATE(CONCAT(SUBSTR(GameData.released_at, 1, 4), '-01-01'))
-                    WHEN GameData.released_at_granularity = 'month' THEN
-                        DATE(CONCAT(SUBSTR(GameData.released_at, 1, 7), '-01'))
-                    ELSE
-                        COALESCE(GameData.released_at, '9999-12-31')
-                END AS normalized_released_at"
-            )
-            ->orderBy('normalized_released_at', $sortDirection);
-    }
-
-    /**
-     * Sort games on the list by the player's progress percentage.
-     * The player's progress percentage is measured by the number of achievements
-     * they've earned divided by the number of achievements published on the game.
-     * If the user has no progress data, their progress is treated as 0.
-     *
-     * If two games have the same progress percentage, a secondary sort is applied
-     * by the number of achievements published.
-     *
-     * @param Builder<Game> $query
-     */
-    private function applyProgressSorting(Builder $query, string $sortDirection = 'asc', ?User $user = null): void
-    {
-        // If there's no user, then we have no progress to sort by. Bail.
-        if (!$user) {
-            $query->orderBy('GameData.sort_title', $sortDirection);
-
-            return;
-        }
-
-        $query
-            ->leftJoin('player_games', function ($join) use ($user) {
-                $join->on('player_games.game_id', '=', 'GameData.ID')
-                    ->where('player_games.user_id', '=', $user->id);
-            })
-            ->selectRaw("
-                CASE
-                    WHEN player_games.completion_percentage IS NULL THEN 0
-                    ELSE player_games.completion_percentage
-                END AS progress_percentage
-            ")
-            ->orderBy('progress_percentage', $sortDirection)
-            ->orderBy('GameData.achievements_published', $sortDirection);
-    }
-
-    /**
-     * Filters games based on whether they have any achievements published.
-     * Possible values are "has", "none", or "either".
-     *
-     * If multiple values are given, only the first one is considered.
-     *
-     * @param Builder<Game> $query
-     */
-    private function applyAchievementsPublishedFilter(Builder $query, array $filterValues): void
-    {
-        // Bail early if necessary. If the user gives both options, it's the "either" case.
-        if (empty($filterValues) || count($filterValues) === 2) {
-            return;
-        }
-
-        // $filterValues is an array, but we only consider a single value.
-        $value = $filterValues[0];
-
-        switch ($value) {
-            case 'has':
-                $query->where('GameData.achievements_published', '>', 0);
-                break;
-
-            case 'none':
-                $query->where(function ($q) {
-                    $q->where('GameData.achievements_published', 0)
-                        ->orWhereNull('GameData.achievements_published');
-                });
-                break;
-
-            case 'either':
-            default:
-                break;
-        }
-    }
-
-    /**
-     * Filter games based on the player's award status.
-     * The player's awards can represent various levels of completion for each game.
-     * The supported awards are beaten (softcore), beaten, completed, and mastered.
-     *
-     * The filtering logic works as follows:
-     *
-     * 1. "unfinished": Filters games where the player has no award of any kind for the game.
-     * 2. "finished": Filters games where the player has any award for the game.
-     * 3. "beaten_softcore": Filters games where the player has a softcore beaten award.
-     * 4. "beaten_hardcore": Filters games where the player has a hardcore beaten award.
-     * 5. "completed": Filters games where the player has a softcore mastery award.
-     * 6. "mastered": Filters games where the player has a hardcore mastery award.
-     *
-     * The SiteAwards.AwardDataExtra field is used to differentiate between softcore (0) and hardcore (1).
-     *
-     * If the user is not provided (ie: they aren't logged in), no filtering will be performed.
-     *
-     * @param Builder<Game> $query
-     */
-    private function applyAwardFilter(Builder $query, array $awardValues, ?User $user = null): void
-    {
-        // If there's no user, then we have no awards to filter by. Bail.
-        if (!$user) {
-            return;
-        }
-
-        /**
-         * We'll pull this data from SiteAwards. Similar data does exist on
-         * player_games, but it is not static. In other words, on player_games,
-         * we revoke the "100% completion" when a set gets revised. This is not
-         * how we communicate the mastery awards UX to players, so we reach for
-         * SiteAwards data instead.
-         */
-        $query->where(function ($query) use ($awardValues, $user) {
-            foreach ($awardValues as $award) {
-                switch ($award) {
-                    /*
-                     * games where the player has no award of any kind for the game
-                     */
-                    case 'unfinished':
-                        $query->orWhereNotExists(function ($subQuery) use ($user) {
-                            $subQuery->select(DB::raw(1))
-                                ->from('SiteAwards')
-                                ->whereColumn('SiteAwards.AwardData', 'GameData.ID')
-                                ->where('SiteAwards.user_id', $user->id)
-                                ->whereIn('SiteAwards.AwardType', [AwardType::GameBeaten, AwardType::Mastery]);
-                        });
-                        break;
-
-                    /*
-                     * games where the player has any award for the game
-                     */
-                    case 'finished':
-                        $query->orWhereExists(function ($subQuery) use ($user) {
-                            $subQuery->select(DB::raw(1))
-                                ->from('SiteAwards')
-                                ->whereColumn('SiteAwards.AwardData', 'GameData.ID')
-                                ->where('SiteAwards.user_id', $user->id)
-                                ->whereIn('SiteAwards.AwardType', [AwardType::GameBeaten, AwardType::Mastery]);
-                        });
-                        break;
-
-                    /*
-                     * games where the player has a softcore beaten award
-                     */
-                    case 'beaten_softcore':
-                        $query->orWhereExists(function ($subQuery) use ($user) {
-                            $subQuery->select(DB::raw(1))
-                                ->from('SiteAwards')
-                                ->whereColumn('SiteAwards.AwardData', 'GameData.ID')
-                                ->where('SiteAwards.user_id', $user->id)
-                                ->where('SiteAwards.AwardType', AwardType::GameBeaten)
-                                ->where('SiteAwards.AwardDataExtra', UnlockMode::Softcore)
-                                ->whereNotExists(function ($excludeSubQuery) use ($user) {
-                                    $excludeSubQuery->select(DB::raw(1))
-                                        ->from('SiteAwards as sa2')
-                                        ->whereColumn('sa2.AwardData', 'GameData.ID')
-                                        ->where('sa2.user_id', $user->id)
-                                        ->where('sa2.AwardType', AwardType::Mastery);
-                                });
-                        });
-                        break;
-
-                    /*
-                     * games where the player has a softcore beaten award
-                     */
-                    case 'beaten_hardcore':
-                        $query->orWhereExists(function ($subQuery) use ($user) {
-                            $subQuery->select(DB::raw(1))
-                                ->from('SiteAwards')
-                                ->whereColumn('SiteAwards.AwardData', 'GameData.ID')
-                                ->where('SiteAwards.user_id', $user->id)
-                                ->where('SiteAwards.AwardType', AwardType::GameBeaten)
-                                ->where('SiteAwards.AwardDataExtra', UnlockMode::Hardcore)
-                                ->whereNotExists(function ($excludeSubQuery) use ($user) {
-                                    $excludeSubQuery->select(DB::raw(1))
-                                        ->from('SiteAwards as sa2')
-                                        ->whereColumn('sa2.AwardData', 'GameData.ID')
-                                        ->where('sa2.user_id', $user->id)
-                                        ->where('sa2.AwardType', AwardType::Mastery);
-                                });
-                        });
-                        break;
-
-                    /*
-                     * games where the player has a softcore mastery award
-                     */
-                    case 'completed':
-                        $query->orWhereExists(function ($subQuery) use ($user) {
-                            $subQuery->select(DB::raw(1))
-                                ->from('SiteAwards')
-                                ->whereColumn('SiteAwards.AwardData', 'GameData.ID')
-                                ->where('SiteAwards.user_id', $user->id)
-                                ->where('SiteAwards.AwardType', AwardType::Mastery)
-                                ->where('SiteAwards.AwardDataExtra', UnlockMode::Softcore)
-                                ->whereNotExists(function ($excludeSubQuery) use ($user) {
-                                    $excludeSubQuery->select(DB::raw(1))
-                                        ->from('SiteAwards as sa2')
-                                        ->whereColumn('sa2.AwardData', 'GameData.ID')
-                                        ->where('sa2.user_id', $user->id)
-                                        ->where('sa2.AwardType', AwardType::Mastery)
-                                        ->where('sa2.AwardDataExtra', UnlockMode::Hardcore);
-                                });
-                        });
-                        break;
-
-                    /*
-                     * games where the player has a hardcore mastery award
-                     */
-                    case 'mastered':
-                        $query->orWhereExists(function ($subQuery) use ($user) {
-                            $subQuery->select(DB::raw(1))
-                                ->from('SiteAwards')
-                                ->whereColumn('SiteAwards.AwardData', 'GameData.ID')
-                                ->where('SiteAwards.user_id', $user->id)
-                                ->where('SiteAwards.AwardType', AwardType::Mastery)
-                                ->where('SiteAwards.AwardDataExtra', UnlockMode::Hardcore);
-                        });
-                        break;
-
-                    default:
-                        break;
-                }
-            }
-        });
     }
 
     /**
