+++ lang/de_DE.json
@@ -129,6 +129,7 @@
     "Emulator": "Emulator",
     "Emulator Core": "Emulator Core",
     "Emulator Version": "Emulator Version",
+    "Ends <1>{{when}}</1>": "Ends <1>{{when}}</1>",
     "Example: The achievement requires collecting 10 coins. You collected 10 coins, but the achievement didn't unlock.": "Example: The achievement requires collecting 10 coins. You collected 10 coins, but the achievement didn't unlock.",
     "Example: The achievement requires collecting 10 coins. You collected 5 coins, and it unlocked.": "Example: The achievement requires collecting 10 coins. You collected 5 coins, and it unlocked.",
     "FAQ": "FAQ",
@@ -150,7 +151,6 @@
     "Go to next page": "Zur nächsten Seite springen",
     "Go to previous page": "Eine Seite zurückspringen",
     "Hardcore": "Hardcore",
-    "Has achievements": "Hat Erfolge",
     "Hash compatibility questions.": "Fragen zur Kompatibilität des Hashes.",
     "Hide": "Verstecke",
     "Hide missable achievement indicators": "Verstecke die Indikatoren für verpassbare Erfolge",
@@ -233,6 +233,7 @@
     "Please write your ticket description in English. This helps ensure it can be resolved as quickly as possible.": "Please write your ticket description in English. This helps ensure it can be resolved as quickly as possible.",
     "Points": "Punkte",
     "Points Earned Since {{date}}": "Erreichte Punkte seit {{date}}",
+    "Prefer absolute dates": "Prefer absolute dates",
     "Preferences": "Präferenzen",
     "Previous": "Vorherige",
     "Previous {{count, number}}": "Vorherige {{count, number}}",
@@ -242,6 +243,7 @@
     "Progress was reset successfully.": "Fortschritt wurde erfolgreich zurückgesetzt.",
     "Proposing a new article idea.": "Eine neue Idee für einen Artikel vorschlagen.",
     "Proposing art updates.": "Zeichen-Updates vorschlagen.",
+    "Published": "Published",
     "Quality Assurance": "Qualitätssicherung",
     "Questions about art-related rule changes.": "Fragen zu kunstbezogenen Regeländerungen.",
     "RAEvents": "RAEvents",
@@ -305,7 +307,6 @@
     "Select an emulator": "Select an emulator",
     "Select an issue": "Select an issue",
     "Settings": "Einstellungen",
-    "Show absolute dates on forum posts": "Absolute Daten in Forumsbeiträge anzeigen",
     "Site Awards": "Website-Auszeichnungen",
     "Site notifications": "Website-Benachrichtigungen",
     "Softcore": "Softcore",
@@ -352,6 +353,7 @@
     "Type / to focus the search field.": "Tippe / um das Suchfeld zu fokussieren.",
     "Type your comment here. Do not post or request any links to copyrighted ROMs.": "Schreibe deinen Kommentar hier. Poste oder fordere keine Links zu urheberrechtlich geschützten ROMs.",
     "Undo": "Rückgängig",
+    "Unpublished": "Unpublished",
     "Unsubscribe": "Abonnent beenden",
     "Unsubscribed!": "Abonnent wurde beendet!",
     "Unsubscribing...": "Abonnent wird beendet...",
@@ -447,5 +449,6 @@
     "{{val, number}} games_one": "{{val, number}} Spiel",
     "{{val, number}} games_other": "{{val, number}} Spiele",
     "{{visible, number}} of {{total, number}} games_one": "{{visible, number}} von {{total, number}} Spiel",
-    "{{visible, number}} of {{total, number}} games_other": "{{visible, number}} von {{total, number}} Spielen"
+    "{{visible, number}} of {{total, number}} games_other": "{{visible, number}} von {{total, number}} Spielen",
+    "Surprise me": "Surprise me"
 }
