+++ lang/fr_FR.json
@@ -151,7 +151,6 @@
     "Go to next page": "Aller à la page suivante",
     "Go to previous page": "Aller à la page précédente",
     "Hardcore": "Hardcore",
-    "Has achievements": "A des succès",
     "Hash compatibility questions.": "Questions sur la compatibilité des empreintes.",
     "Hide": "Masquer",
     "Hide missable achievement indicators": "Masquer les indicateurs de succès manquables",
@@ -456,5 +455,6 @@
     "{{val, number}} games_other": "{{val, number}} jeux",
     "{{visible, number}} of {{total, number}} games_one": "{{visible, number}} sur {{total, number}} jeu",
     "{{visible, number}} of {{total, number}} games_many": "{{visible, number}} of {{total, number}} jeux",
-    "{{visible, number}} of {{total, number}} games_other": "{{visible, number}} sur {{total, number}} jeux"
+    "{{visible, number}} of {{total, number}} games_other": "{{visible, number}} sur {{total, number}} jeux",
+    "Surprise me": "Surprise me"
 }
