+++ lang/pt_BR.json
@@ -151,7 +151,6 @@
     "Go to next page": "Ir para a próxima página",
     "Go to previous page": "Ir para a página anterior",
     "Hardcore": "Hardcore",
-    "Has achievements": "Possui conquistas",
     "Hash compatibility questions.": "Perguntas sobre compatibilidade de hash.",
     "Hide": "Ocultar",
     "Hide missable achievement indicators": "Ocultar indicadores de conquistas perdidas",
@@ -456,5 +455,6 @@
     "{{val, number}} games_other": "{{val, number}} jogos",
     "{{visible, number}} of {{total, number}} games_one": "{{visible, number}} de {{total, number}} jogo",
     "{{visible, number}} of {{total, number}} games_many": "{{visible, number}} of {{total, number}} jogos",
-    "{{visible, number}} of {{total, number}} games_other": "{{visible, number}} de {{total, number}} jogos"
+    "{{visible, number}} of {{total, number}} games_other": "{{visible, number}} de {{total, number}} jogos",
+    "Surprise me": "Surprise me"
 }
