+++ resources/js/features/game-list/components/DataTableToolbar/DataTableDesktopToolbar.tsx
@@ -10,12 +10,14 @@ import { DataTableSearchInput } from '../DataTableSearchInput';
 import { DataTableViewOptions } from '../DataTableViewOptions';
 import { DataTableAchievementsPublishedFilter } from './DataTableAchievementsPublishedFilter';
 import { DataTableSystemFilter } from './DataTableSystemFilter';
+import { RandomGameButton } from './RandomGameButton';
 
 interface DataTableDesktopToolbarProps<TData> {
   table: Table<TData>;
   unfilteredTotal: number | null;
 
   defaultColumnFilters?: ColumnFiltersState;
+  randomGameApiRouteName?: RouteName;
   tableApiRouteName?: RouteName;
 }
 
@@ -23,6 +25,7 @@ export function DataTableDesktopToolbar<TData>({
   table,
   unfilteredTotal,
   defaultColumnFilters = [],
+  randomGameApiRouteName = 'api.game.random',
   tableApiRouteName = 'api.game.index',
 }: DataTableDesktopToolbarProps<TData>) {
   const { filterableSystemOptions } = usePageProps<{
@@ -57,7 +60,7 @@ export function DataTableDesktopToolbar<TData>({
       </div>
 
       <div className="flex items-center justify-between gap-3 md:justify-normal">
-        <p className="text-neutral-200 light:text-neutral-900">
+        <p className="mr-2 text-neutral-200 light:text-neutral-900">
           {unfilteredTotal && unfilteredTotal !== table.options.rowCount ? (
             <>
               {t('{{visible, number}} of {{total, number}} games', {
@@ -76,7 +79,15 @@ export function DataTableDesktopToolbar<TData>({
           )}
         </p>
 
-        <DataTableViewOptions table={table} />
+        <div className="flex items-center gap-3 md:gap-1">
+          <RandomGameButton
+            variant="toolbar"
+            apiRouteName={randomGameApiRouteName}
+            columnFilters={currentFilters}
+          />
+
+          <DataTableViewOptions table={table} />
+        </div>
       </div>
     </div>
   );
