+++ resources/js/features/game-list/components/DataTableToolbar/RandomGameButton/index.ts
@@ -1,0 +1,1 @@
+export * from './RandomGameButton';
