+++ resources/js/features/game-list/components/WantToPlayGamesMainRoot/WantToPlayGamesMainRoot.test.tsx
@@ -15,6 +15,8 @@ import {
 
 import { WantToPlayGamesMainRoot } from './WantToPlayGamesMainRoot';
 
+vi.mock('../DataTableToolbar/RandomGameButton');
+
 // Suppress AggregateError invocations from unmocked fetch calls to the back-end.
 console.error = vi.fn();
