+++ resources/js/ziggy.d.ts
@@ -206,6 +206,7 @@ declare module 'ziggy-js' {
     "claims.active": [],
     "pulse": [],
     "api.game.index": [],
+    "api.game.random": [],
     "game.hashes.index": [
         {
             "name": "game",
@@ -370,6 +371,7 @@ declare module 'ziggy-js' {
         }
     ],
     "api.user-game-list.index": [],
+    "api.user-game-list.random": [],
     "api.user-game-list.store": [
         {
             "name": "game",
